Don't need to makeGameFolder now, by default the symlinks of the asset folders will be created instead of copying them over
Only when makeGameFolder.sh is called will the assets be actually copied over.

+++ makeGameFolder.sh
@@ -1,46 +1,32 @@
 set -e
 cd "$(dirname "${0}")/.."
 
+PLATFORM=${1-5}
+if [[ $PLATFORM != 1 ]] && [[ $PLATFORM != 5 ]]; then
+	echo "Usage: 1 for Linux, 5 for XCompiling for Windows (Default)"
+	exit 1
+fi
+
 mkdir -p output
 cd output
 
-if [[ $# == 0 ]]; then folderList="animations categories ground music objects sounds sprites transitions"; fi
-if [[ $# > 0 ]]; then folderList="$@"; fi
+folderList="animations categories ground music objects sounds sprites transitions"
 
 #Copying from OneLifeData7
 for f in $folderList; do
     
 	###### Create sym link only
-	# if [ ! -h $f ]; then ln -s ../OneLifeData7/$f .; fi
+	# if [[ $PLATFORM == 1 ]] && [ ! -h $f ]; then ln -s ../OneLifeData7/$f .; fi
+    # if [[ $PLATFORM == 5 ]] && [ ! -h $f ]; then ../miniOneLifeCompile/mklink.sh /J $f ../OneLifeData7/$f; fi
 	
 	###### Full copy from OneLifeData7 repo
 	# if [ ! -d $f ]; then cp -RL -v ../OneLifeData7/$f .; fi
 	
 	###### Sync only when different
-	if [ -h $f ]; then rm $f; fi
-	mkdir -p $f
-	cd $f
-	rsync -vr ../../OneLifeData7/$f/ .
-	cd ..
+    if [ -h $f ]; then rm $f; fi
+    mkdir -p $f
+    cd $f
+    rsync -vr ../../OneLifeData7/$f/ .
+    cd ..
 	
 done;
-
-if [[ $# == 0 ]]; then
-	cp ../OneLifeData7/dataVersionNumber.txt .
-
-	#Copying from OneLife
-	cp ../OneLife/documentation/Readme.txt .
-	cp ../OneLife/no_copyright.txt .
-	cp ../OneLife/gameSource/language.txt .
-	cp ../OneLife/gameSource/us_english_60.txt .
-	cp ../OneLife/gameSource/reverbImpulseResponse.aiff .
-	cp ../OneLife/gameSource/wordList.txt .
-	cp -r ../OneLife/gameSource/graphics .
-	cp -r ../OneLife/gameSource/otherSounds .
-	cp -r ../OneLife/gameSource/settings .
-	cp -r ../OneLife/gameSource/languages .
-
-
-	#missing SDL.dll
-	cp ../OneLife/build/win32/SDL.dll .
-fi;
